풀하기 위한 커밋

Merge branch 'master' of https://github.com/sck3668/adaco.git

Conflicts:
	adaco/target/classes/com/icia/adaco/dao/ArtDao.class

+++ adaco/src/main/java/com/icia/adaco/entity/Art.java
@@ -23,7 +23,7 @@ public class Art {
 	private String returnAddress;
 	private String courier;
 	private Integer readCnt;
-	private Integer accumulated; // 택배사
+	private Double accumulated; // 택배사
 	private Integer favoriteCnt;
 	private Boolean isFavorite;
 	private String category;
